Changes to the header
The changes include removing the image that was there and used the css
code that was previously there and made adjustments to them. Also added
the fade in function to the header text also.

Added the jquery plugin for a sticky scroll
-we may not keep this because of the functionailty is not working
correctly on my site. More testing to come.